[SYCL] Initial changes for the second version of sycl_ext_oneapi_group_sort extension
Current implementation is aligned with the first version of the
extension: https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/experimental/sycl_ext_oneapi_group_sort.asciidoc

This second version of the extension:
https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/proposed/sycl_ext_oneapi_group_sort.asciidoc
changes the API by introducing separate sorter objects for
sort_over_group and joint_sort.

Even though extension is experimental and it is formally allowed to
break API any time, we want to be cautios and save sorters provided by
the first version of the API for some period of time to give users an
opportunity to move to the second version of API.

Test was changed to test both versions.